Don't assume sizeof(void *) == sizeof(unsigned long).

+++ etc/afpd/desktop.c
@@ -697,7 +697,7 @@ int afp_addcomment(AFPObj *obj _U_, char *ibuf, size_t ibuflen _U_, char *rbuf _
 	return get_afp_errno(AFPERR_NOOBJ);
     }
 
-    if ((u_long)ibuf & 1 ) {
+    if ((intptr_t)ibuf & 1 ) {
         ibuf++;
     }
